Moved most images to 'common/images'
