--- src/components/Menu_Management/Zmones_Tab/PeopleListComponents/images/SvgDownload.jsx
@@ -1,18 +1,0 @@
-import React, { Component } from 'react';
-export default () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="16"
-    height="15"
-    viewBox="0 0 16 15"
-  >
-    <g fill="none" fill-rule="evenodd">
-      <path d="M-2-2h20v20H-2z" />
-      <path
-        fill="#FFF"
-        fill-rule="nonzero"
-        d="M15.73 10.273a.91.91 0 0 1 .27.665v3.125a.91.91 0 0 1-.27.664.883.883 0 0 1-.653.273H.923a.883.883 0 0 1-.654-.273.911.911 0 0 1-.269-.664v-3.125c0-.26.09-.482.27-.665A.884.884 0 0 1 .922 10h4.471l1.298 1.328c.372.365.808.547 1.308.547s.936-.182 1.308-.547L10.615 10h4.462c.256 0 .474.091.654.273zm-3.605 3.292a.606.606 0 0 0 .183-.44c0-.17-.061-.316-.183-.44a.587.587 0 0 0-.433-.185.587.587 0 0 0-.432.186.605.605 0 0 0-.183.44c0 .168.06.315.183.439a.587.587 0 0 0 .432.185.588.588 0 0 0 .433-.185zm2.462 0a.605.605 0 0 0 .182-.44.605.605 0 0 0-.182-.44.587.587 0 0 0-.433-.185.587.587 0 0 0-.433.186.605.605 0 0 0-.182.44c0 .168.06.315.182.439a.587.587 0 0 0 .433.185.587.587 0 0 0 .433-.185zM7.56 10.808l-4.366-4.53c-.201-.195-.247-.43-.136-.707.11-.263.302-.394.575-.394h2.495V.647c0-.175.061-.327.185-.455A.59.59 0 0 1 6.752 0h2.496a.59.59 0 0 1 .438.192c.123.128.185.28.185.455v4.53h2.495c.273 0 .465.131.575.394.11.276.065.512-.136.708l-4.366 4.529A.57.57 0 0 1 8 11a.57.57 0 0 1-.439-.192z"
-      />
-    </g>
-  </svg>
-);
