--- src/components/Menu_Management/Zmones_Tab/PeopleListComponents/images/TableSearch.jsx
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-import SvgSearch from './SvgSearch.jsx';
-
-const TableSearch = () => {
-  return (
-    <div className="table__search-icon">
-      <SvgSearch />
-    </div>
-  );
-};
-
-export default TableSearch;
