--- src/components/SelfProfile/tabs/MyTravelsTab/TravelListComponents/images/SvgCaretLines.jsx
@@ -1,19 +1,0 @@
-import React, { Component } from 'react';
-export default () => (
-  <svg
-    className="location-caret"
-    xmlns="http://www.w3.org/2000/svg"
-    width="12"
-    height="8"
-    viewBox="0 0 12 8"
-  >
-    <g fill="none" fill-rule="evenodd">
-      <path
-        fill="#1D1E1F"
-        fill-rule="nonzero"
-        d="M4.667 8h2.666V6.667H4.667V8zM0 0v1.333h12V0H0zm2 4.667h8V3.333H2v1.334z"
-      />
-      <path d="M-2-4h16v16H-2z" />
-    </g>
-  </svg>
-);
