--- src/components/SelfProfile/tabs/MyTravelsTab/TravelListComponents/images/SvgHand.jsx
@@ -1,18 +1,0 @@
-import React, { Component } from 'react';
-export default () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="24"
-    height="20"
-    viewBox="0 0 20 16"
-  >
-    <g fill="none" fill-rule="evenodd">
-      <path d="M0-1h20v20H0z" />
-      <path
-        fill="#8995A4"
-        fill-rule="nonzero"
-        d="M.833 14.3c1.065.32 4.775 1.367 6.142 1.367h4.746c.64 0 1.16-.5 1.16-1.114 0-.614-.52-1.113-1.16-1.113H9.697a.213.213 0 0 1-.215-.21c0-.116.096-.21.215-.21h3.163c.64 0 1.16-.5 1.16-1.113 0-.614-.52-1.114-1.16-1.114h-2.255a.213.213 0 0 1-.215-.21c0-.115.096-.21.215-.21h3.163c.64 0 1.16-.499 1.16-1.113s-.52-1.113-1.16-1.113h-2.255a.213.213 0 0 1-.216-.21c0-.116.097-.21.216-.21h6.494c.64 0 1.16-.5 1.16-1.113 0-.614-.52-1.114-1.16-1.114H7.95a.216.216 0 0 1-.202-.136.208.208 0 0 1 .06-.231l3.069-2.609.032-.03c.204-.21.317-.484.318-.772v-.008a1.097 1.097 0 0 0-.331-.782c-.264-.256-.77-.458-1.676.044-1.375.76-2.68 1.593-3.835 2.33-.86.548-1.67 1.066-2.355 1.452-.438.247-1.588.761-2.198 1.046V14.3z"
-      />
-    </g>
-  </svg>
-);
