--- src/components/SelfProfile/tabs/MyTravelsTab/TravelListComponents/images/SvgReorder.jsx
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-export default () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="8"
-    height="12"
-    viewBox="0 0 8 12"
-  >
-    <g fill="#DBDBE3" fill-rule="evenodd">
-      <path d="M.25 7.5L4 11.25 7.75 7.5zM.25 4.25L4 .5l3.75 3.75z" />
-    </g>
-  </svg>
-);
