--- src/components/SelfProfile/tabs/MyTravelsTab/TravelListComponents/images/SvgSearch.jsx
@@ -1,18 +1,0 @@
-import React, { Component } from 'react';
-export default () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="15"
-    height="15"
-    viewBox="0 0 17 17"
-  >
-    <g fill="none" fill-rule="evenodd">
-      <path
-        fill="#78869F"
-        fill-rule="nonzero"
-        d="M12.15 10.692h-.768l-.272-.263a6.29 6.29 0 0 0 1.526-4.111 6.318 6.318 0 1 0-6.318 6.318 6.29 6.29 0 0 0 4.111-1.526l.263.272v.768l4.86 4.85L17 15.552l-4.85-4.86zm-5.832 0a4.368 4.368 0 0 1-4.374-4.374 4.368 4.368 0 0 1 4.374-4.374 4.368 4.368 0 0 1 4.374 4.374 4.368 4.368 0 0 1-4.374 4.374z"
-      />
-      <path d="M-3-4h24v24H-3z" />
-    </g>
-  </svg>
-);
