--- src/components/SidebarComp/Sidebar/SidebarComponents/images/SvgProfile.jsx
@@ -1,23 +1,0 @@
-import React from 'react';
-
-const SvgProfile = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      viewBox="0 0 16 16"
-    >
-      <g fill="none" fillRule="evenodd">
-        <path
-          fill="#8995A4"
-          fillRule="nonzero"
-          d="M8 1.9a2.1 2.1 0 1 1 0 4.2 2.1 2.1 0 0 1 0-4.2zm0 9c2.97 0 6.1 1.46 6.1 2.1v1.1H1.9V13c0-.64 3.13-2.1 6.1-2.1zM8 0C5.79 0 4 1.79 4 4s1.79 4 4 4 4-1.79 4-4-1.79-4-4-4zm0 9c-2.67 0-8 1.34-8 4v3h16v-3c0-2.66-5.33-4-8-4z"
-        />
-        <path d="M-4-4h24v24H-4z" />
-      </g>
-    </svg>
-  );
-};
-
-export default SvgProfile;
