--- src/components/SidebarComp/SidebarAdmin/SidebarAdminComponents/images/SvgAdmin.jsx
@@ -1,23 +1,0 @@
-import React from 'react';
-
-const SvgAdmin = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="22"
-      height="16"
-      viewBox="0 0 22 16"
-    >
-      <g fill="none" fillRule="evenodd">
-        <path
-          fill="#8995A4"
-          fillRule="nonzero"
-          d="M18 4l-4 4h3c0 3.31-2.69 6-6 6a5.87 5.87 0 0 1-2.8-.7l-1.46 1.46A7.93 7.93 0 0 0 11 16c4.42 0 8-3.58 8-8h3l-4-4zM5 8c0-3.31 2.69-6 6-6 1.01 0 1.97.25 2.8.7l1.46-1.46A7.93 7.93 0 0 0 11 0C6.58 0 3 3.58 3 8H0l4 4 4-4H5z"
-        />
-        <path d="M-1-4h24v24H-1z" />
-      </g>
-    </svg>
-  );
-};
-
-export default SvgAdmin;
