--- src/components/SidebarComp/SidebarAdmin/SidebarAdminComponents/images/SvgList.jsx
@@ -1,23 +1,0 @@
-import React from 'react';
-
-const SvgList = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="19"
-      height="16"
-      viewBox="0 0 19 16"
-    >
-      <g fill="none" fillRule="evenodd">
-        <path
-          fill="#8995A4"
-          fillRule="nonzero"
-          d="M2 6.5C1.17 6.5.5 7.17.5 8S1.17 9.5 2 9.5 3.5 8.83 3.5 8 2.83 6.5 2 6.5zm0-6C1.17.5.5 1.17.5 2S1.17 3.5 2 3.5 3.5 2.83 3.5 2 2.83.5 2 .5zm0 12c-.83 0-1.5.68-1.5 1.5s.68 1.5 1.5 1.5 1.5-.68 1.5-1.5-.67-1.5-1.5-1.5zM5 15h14v-2H5v2zm0-6h14V7H5v2zm0-8v2h14V1H5z"
-        />
-        <path d="M-2-4h24v24H-2z" />
-      </g>
-    </svg>
-  );
-};
-
-export default SvgList;
