--- src/components/SidebarComp/SidebarAdmin/SidebarAdminComponents/images/SvgLogo.jsx
@@ -1,20 +1,0 @@
-import React, { Component } from 'react';
-
-const SvgLogo = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="30"
-      height="30"
-      viewBox="0 0 30 30"
-    >
-      <path
-        fill="#EE3957"
-        fillRule="evenodd"
-        d="M0 30V0h30v30H0zm20.635-12.873a2.26 2.26 0 0 1-2.258-2.26 2.257 2.257 0 1 1 4.516 0 2.26 2.26 0 0 1-2.258 2.26zm0-7.344c-.815 0-1.579.196-2.258.534V5.265h-2.825v9.601a5.084 5.084 0 1 0 5.083-5.083zm-11.27 11.21a2.259 2.259 0 1 1 .002-4.519 2.259 2.259 0 0 1-.001 4.518zm4.892-3.629c-.018-.065-.034-.13-.055-.193a6.269 6.269 0 0 1-.222-1.634V9.11h-2.824v4.867a5.083 5.083 0 1 0-1.79 9.837 5.085 5.085 0 0 0 4.183-2.198 9.087 9.087 0 0 0 3.12 2.198l1.414-2.455v-.006a6.188 6.188 0 0 1-3.826-3.99z"
-      />
-    </svg>
-  );
-};
-
-export default SvgLogo;
