--- src/components/SidebarComp/SidebarAdmin/SidebarAdminComponents/images/SvgSchool.jsx
@@ -1,23 +1,0 @@
-import React from 'react';
-
-const SvgSchool = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="12"
-      height="10"
-      viewBox="0 0 12 10"
-    >
-      <g fill="none" fillRule="evenodd">
-        <path d="M0-1h12v12H0z" />
-        <path
-          fill="#00B488"
-          fillRule="nonzero"
-          d="M2.5 5.59v2L6 9.5l3.5-1.91v-2L6 7.5 2.5 5.59zM6 .5l-5.5 3 5.5 3 4.5-2.455V7.5h1v-4L6 .5z"
-        />
-      </g>
-    </svg>
-  );
-};
-
-export default SvgSchool;
