+++ src/components/common/images/SvgDelete.jsx
