+++ src/components/common/images/SvgHand.jsx
