+++ src/components/common/images/SvgLogo.jsx
