+++ src/components/common/images/SvgSchool.jsx
