+++ src/components/common/images/SvgSearch.jsx
